move search page comps into compositions folder
